requirements.txt
beautifulsoup4==4.8.2
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
DoubleMetaphone==0.1
emoji==0.5.4
en-core-web-sm==2.2.5
future==0.18.2
fuzzywuzzy==0.17.0
idna==2.8
IMDbPY==6.8
lxml==4.5.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.1
plac==1.1.3
preshed==3.0.2
probableparsing==0.0.1
probablepeople==0.5.4
python-crfsuite==0.9.6
python-usda==0.4.6
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
spacy==2.2.3
SQLAlchemy==1.3.13
srsly==1.0.1
thinc==7.3.1
tmdbv3api==1.5.1
tqdm==4.42.0
urllib3==1.25.8
wasabi==0.6.0
wincertstore==0.2
